% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ferries.R
\name{add_ferries}
\alias{add_ferries}
\title{Adding ferry-related adjacencies}
\usage{
add_ferries(pref)
}
\arguments{
\item{pref}{sf object of cleaned, collated, census data for a prefecture}
}
\value{
array with two columns, containing start port location to end port location for boats
}
\description{
Adding ferry-related adjacencies
}
\concept{organizedata}
