% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch03CTT.R
\name{AlphaCoefficient}
\alias{AlphaCoefficient}
\title{Alpha Coefficient}
\usage{
AlphaCoefficient(x, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{x}{This should be a data matrix or a Covariance/Phi/Tetrachoric matrix.}

\item{na}{This parameter identifies the numbers or characters that should be treated as missing values when 'x' is a data matrix.}

\item{Z}{This parameter represents a missing indicator matrix. It is only needed if 'x' is a data matrix.}

\item{w}{This parameter is an item weight vector. It is only required if 'x' is a data matrix.}
}
\description{
This function computes Tau-Equivalent Measurement, also known as Cronbach's alpha coefficient, for a given data set.
}
\references{
Cronbach, L. J. (1951). Coefficient alpha and the internal structure of a test. Psychometrika, 16,297–334.
}
