% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch03CTT.R
\name{AlphaIfDel}
\alias{AlphaIfDel}
\title{Alpha Coefficient if Item removed}
\usage{
AlphaIfDel(x, delItem = NULL, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{x}{This should be a data matrix or a Covariance/Phi/Tetrachoric matrix.}

\item{delItem}{Specify the item to be deleted. If NULL, calculations are performed for all cases.}

\item{na}{This parameter identifies the numbers or characters that should be treated as missing values when 'x' is a data matrix.}

\item{Z}{This parameter represents a missing indicator matrix. It is only needed if 'x' is a data matrix.}

\item{w}{This parameter is an item weight vector. It is only required if 'x' is a data matrix.}
}
\description{
This function returns the alpha coefficient when the specified item is excluded.
}
