% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoordinateCleaner-package.R
\docType{data}
\name{buffland}
\alias{buffland}
\title{Global Coastlines buffered by 1 degree}
\source{
\url{http://www.naturalearthdata.com/downloads/10m-physical-vectors/}
}
\description{
A \code{SpatialPolygonsDataFrame} with global coastlines, with a 1 degree buffer to extent coastlines as alternative reference for \code{\link{cc_sea}}. Can be useful to identify species in the sea, without flagging records in mangroves, marshes, etc.
}
\examples{

data("buffland")
}
\keyword{gazetteers}
