% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_data_motto.R
\name{build_data_motto}
\alias{build_data_motto}
\title{Build Data Motto Website}
\usage{
build_data_motto()
}
\description{
Build Data Motto Website
}
\details{
It re-generate a \code{posts.json}, update the \code{index.html}, and create the whole
website based on the updated \code{posts} folder.
}
