% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_docs.R
\name{build_docs}
\alias{build_docs}
\title{build_docs}
\usage{
build_docs(force = FALSE)
}
\arguments{
\item{force}{If TRUE, rebuild everything}
}
\description{
Build documentation for R packages for for docs.evanodell.com
}
