% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_kokumin.R
\name{calc_kokumin}
\alias{calc_kokumin}
\title{Modify Column of Prefecture Shapefile with the Population of Japanese Nationals
Note: the Japanese census data randomly merges some areas together in their data collection process,
so there are some inconsistencies (i.e. areas with '0' population)}
\usage{
calc_kokumin(pref, age_pops)
}
\arguments{
\item{pref}{an sf object with cleaned, collated version of the census data}

\item{age_pops}{csv containing population by age + foreigner population}
}
\value{
a dataframe similar to pref, with an additional KOKUMIN column containing the number of Japanese nationals
}
\description{
Modify Column of Prefecture Shapefile with the Population of Japanese Nationals
Note: the Japanese census data randomly merges some areas together in their data collection process,
so there are some inconsistencies (i.e. areas with '0' population)
}
\concept{calculate}
