% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{CalcHPModel}
\alias{CalcHPModel}
\title{A Function to calculate the Hydropriming model parameters.}
\usage{
CalcHPModel()
}
\description{
This function calculates the minimal water potential for priming effects (Psib50min).
}
\examples{
CalcHPModel()
CalcHPModel()
}
\keyword{hydropriming}
\keyword{model}
\keyword{parameters}
