% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{CalcHTModel}
\alias{CalcHTModel}
\title{A Function to calculate the Hydrotime model parameters.}
\usage{
CalcHTModel()
}
\description{
This function calculates the hydrotime constant (HT), the median water potential base (Psib50) and the standard deviation (sigma).
}
\examples{
CalcHTModel()
CalcHTModel()
}
\keyword{hydrotime}
\keyword{model}
\keyword{parameters}
