% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{CalcHTPModel}
\alias{CalcHTPModel}
\title{A Function to calculate the Hydrothermal priming model parameters.}
\usage{
CalcHTPModel()
}
\description{
This function calculates the minimal water potential for priming effects (Psibmin50) and minimal temperature (Tmin).
}
\examples{
CalcHTPModel()
CalcHTPModel()
}
\keyword{hydrothermal}
\keyword{model}
\keyword{parameters}
\keyword{priming}
