% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{CalcHTTModel}
\alias{CalcHTTModel}
\title{A Function to calculate the Hydrothermal time model parameters.}
\usage{
CalcHTTModel()
}
\description{
This function calculates the hydrohermal time constant (HT), the temperature base (tb), the median water potential base (Psib50) and the standard deviation (sigma).
}
\examples{
CalcHTTModel()
CalcHTTModel()
}
\keyword{hydrothermal}
\keyword{model}
\keyword{parameters}
\keyword{time}
