% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{CalcHTTModelSupra}
\alias{CalcHTTModelSupra}
\title{A Function to calculate the Supra-optimal hydrothermal time model parameters.}
\usage{
CalcHTTModelSupra()
}
\description{
This function calculates the supra-optimal hydrothermal time constant (HT), the optimal temperature (To), the median water potential base (Psib50), the standard deviation (sigma) and the slope of the relationship between Psib50 and temperature (kt).
}
\examples{
CalcHTTModelSupra()
CalcHTTModelSupra()

}
\keyword{hydrothermal}
\keyword{model}
\keyword{parameters}
\keyword{supra-optimal}
\keyword{time}
