% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{CALCnPLOTModel}
\alias{CALCnPLOTModel}
\title{A Function to calculate and plot the PBTM selected.}
\usage{
CALCnPLOTModel()
}
\arguments{
\item{CALCnPLOTModel()}{calculates the selected model}
}
\description{
This function triggers the model calculation and outputs calculated parameters and graph prediction.
}
\keyword{PBTM}
