% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{CalcSupraHTModel}
\alias{CalcSupraHTModel}
\title{A Function to calculate the supra-optimal Hydrotime model parameters.}
\usage{
CalcSupraHTModel()
}
\description{
This function calculates the supra-optimal hydrotime constant (HT), the median water potential base (Psib50) and the variation (sigma).
}
\examples{
CalcSupraHTModel()
CalcSupraHTModel()
}
\keyword{hydrotime}
\keyword{model}
\keyword{parameters}
\keyword{supra-optimal}
