% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{CalcTTSubOModel}
\alias{CalcTTSubOModel}
\title{A Function to calculate the Thermaltime model parameters.}
\usage{
CalcTTSubOModel()
}
\description{
This function calculates the temperature base (Tb), the thermal time to 50% of the population (ThetaT50) and the standard deviation (sigma).
}
\examples{
CalcTTSubOModel()
CalcTTSubOModel()

}
\keyword{Thermal}
\keyword{model}
\keyword{parameters}
\keyword{time}
