% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauda-package.R
\docType{package}
\name{cauda-package}
\alias{cauda-package}
\title{cauda: Causal Automated Unified Data Analysis}
\description{
A simple, high-level toolkit for exploratory causal data analysis. Provides automated causal discovery, graph editing, correlation networks, partial dependence plots, decision optimization, and conditional independence testing for tabular data frames.
}
\author{
\strong{Maintainer}: Aadi Soni \email{aadipsoni@gmail.com}

}
\keyword{internal}
