% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauda-docs.R
\name{cauda.add}
\alias{cauda.add}
\title{Add an arrow to the causal DAG}
\usage{
cauda.add(dag, x, y, highlight = NULL)
}
\arguments{
\item{dag}{A \code{bn} object returned by \code{cauda.dag()}.}

\item{x}{Character. The source node name.}

\item{y}{Character. The destination node name.}

\item{highlight}{Character. Column name to highlight in red. Default is NULL.}
}
\value{
Invisibly returns the updated \code{bn} object.
}
\description{
Adds a directed arrow from node x to node y and replots the DAG.
}
\examples{
df_ready <- cauda.prep(mtcars, verbose = FALSE)
dag <- cauda.dag(df_ready, verbose = FALSE)
dag <- cauda.add(dag, "wt", "qsec")

}
