% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauda-docs.R
\name{cauda.flip}
\alias{cauda.flip}
\title{Flip the direction of an arrow in the causal DAG}
\usage{
cauda.flip(dag, x, y, highlight = NULL)
}
\arguments{
\item{dag}{A \code{bn} object returned by \code{cauda.dag()}.}

\item{x}{Character. One endpoint of the arrow.}

\item{y}{Character. The other endpoint of the arrow.}

\item{highlight}{Character. Column name to highlight in red. Default is NULL.}
}
\value{
Invisibly returns the updated \code{bn} object.
}
\description{
Reverses the direction of the arrow between nodes x and y and replots the DAG.
}
\examples{
df_ready <- cauda.prep(mtcars, verbose = FALSE)
dag <- cauda.dag(df_ready, verbose = FALSE)
dag <- cauda.flip(dag, "cyl", "disp")

}
