% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkstorm.R
\name{checkstorm}
\alias{checkstorm}
\title{Internal check to verify if input data is properly formatted}
\usage{
checkstorm(dta, agncy)
}
\arguments{
\item{dta}{to check formatting on. Should be the result of \code{get_storms()}}
}
\value{
a single data frame that is the same as the entered into parent function, or one of the listed data frames entered into the parent function.
}
\description{
Internal check to verify if input data is properly formatted
}
