% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_2020_census.R
\name{clean_2020_census}
\alias{clean_2020_census}
\title{Clean latest 2020 data frame}
\usage{
clean_2020_census(total, foreigner)
}
\arguments{
\item{total}{downloaded data frame of the total population of 2020 census}

\item{foreigner}{downloaded data frame of the foreigner population data of 2020 census}
}
\value{
a data frame object of combined and cleaned 2020 census data
}
\description{
Clean latest 2020 data frame
}
\concept{getdata}
