% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_jcdf.R
\name{clean_jcdf}
\alias{clean_jcdf}
\title{Download JCDF (Japan Census Data File)}
\usage{
clean_jcdf(pref_raw)
}
\arguments{
\item{pref_raw}{raw prefecture shapefile data downloaded using download_shp function}
}
\value{
a sf object with cleaned, collated version of the census data
}
\description{
Download JCDF (Japan Census Data File)
}
\concept{getdata}
