% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{CleanGermData}
\alias{CleanGermData}
\title{A Function to clean cumulative curves on dataset.}
\usage{
CleanGermData()
}
\arguments{
\item{CleanGermData()}{will remove repetitive points}
}
\description{
This function removes repetitive percentages, keeping only the intial presence of the value
}
\examples{
CleanGermData()
CleanGermData()
}
\keyword{Clean}
\keyword{germination}
\keyword{percentage}
\keyword{repetitive}
