% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_config_Dotto.R
\name{Config_Dotto_to_json}
\alias{Config_Dotto_to_json}
\title{Create a config json variable based on each dots}
\usage{
Config_Dotto_to_json(path)
}
\description{
Each \code{Dot} will be evaluated and the html output will be saved as a separated object
that can be sent to Dotto creation funtion.
}
\details{
It first renders the Dotto in a simple \code{html_document} format and extract the html codes
related to each chunk (\code{dot}). The output html code, attached with dot number, part, and language
will be saved as a json object and will be passed to Dotto creation function for applying the style.
}
