% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_posts.R
\name{config_posts}
\alias{config_posts}
\title{Config all posts}
\usage{
config_posts(path)
}
\description{
It creates a \code{posts.json} file in the root the package with
all the high level information (\code{metadata}) of posts.

@importFrom here here
@import RJSONIO
@import rmarkdown
@importFrom glue glue
}
