% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{convert_deltaZ2_matrix}
\alias{convert_deltaZ2_matrix}
\title{convert_deltaZ2_matrix}
\usage{
convert_deltaZ2_matrix(
  df,
  use_row.col,
  use_column.col,
  deltaZ.col,
  use_method = "max"
)
}
\arguments{
\item{df}{a data frame with deltaZ scores,  it could be created by run \code{generate.masterTable.SINBA }.}

\item{use_row.col}{column name in the df to be used as rows in the output matrix.}

\item{use_column.col}{colum name in the df to be used as columns in the output matrix.}

\item{deltaZ.col}{column name with the deltaZ scores, the values will be used in the output matrix.}

\item{use_method}{the method used to aggregate the deltaZ values from same seed/partner drivers. Options: "max", "mean", "min" and "median". default="max"}
}
\description{
Convert a dataframe with selected columns(including the deltaZ score column) to a matrix.
}
\examples{

ms_tab<-SINBA.par$ms_tab
df<-convert_deltaZ2_matrix(df=ms_tab,use_row.col="SINBA.seed_geneSymbol",use_column.col="SINBA.partner_geneSymbol",deltaZ.col="delta_Z.G3.Vs.others_SINBA",use_method="max")
\dontrun{
}
}
