% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\docType{data}
\name{count_obs}
\alias{count_obs}
\title{Count data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 85 rows and 26 columns.
}
\usage{
count_obs
}
\description{
Count data
}
\keyword{datasets}
