% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_create_Dotto.R
\name{create_Dotto}
\alias{create_Dotto}
\title{Create a new Dotto}
\usage{
create_Dotto(title = NULL, lang = "r", open = T)
}
\arguments{
\item{title}{Title of the dotto post.}

\item{lang}{The programming language for writing the Dotto. Default is \code{R}. Other
options are \code{python}, \code{julia}, \code{sql}, \code{bash}, \code{js}, \code{node}, \code{d3}, \code{Rcpp}, \code{stan}. It is not
case-sensitive.}

\item{open}{If open the dotto file in Rstudio or not. Default is \code{TRUE}.}
}
\description{
It creates a \code{Dotto} (.Rmd) template based on a title and engine language.
}
\details{
For the \code{yml} metadata, here are more details about some tags:
\itemize{
\item \code{lang}: \code{R}, \code{python}, \code{julia}, \code{sql}, \code{bash}, \code{node}, etc.
\item \code{categories}: \verb{Importing", "Cleaning}, \code{Manipulation}, \code{Visualization}, \code{Machine-Learning}, \code{Statistical-Learning}, \code{Reporting}, \code{Data-Engineering}
\item \code{site}: A personal url address, e.g., a personal website, twitter account, linkedin, etc.
\item \code{img_url}: A canonical link into your img that should be saved under \verb{_files} dir.
\item \code{applications}: Sector specific application require domain knowledge expertise like \code{Finance}, \code{Healthcare}, \code{Insurance}, etc.
\item \code{cover_image}: An external/internal link to the img that will be displayed in the cards.
}
}
