% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_create_family.R
\name{create_family}
\alias{create_family}
\title{Create a \code{Dotto} family be generating a new \code{Dotto} based on another existing \code{Dotto}.}
\usage{
create_family(open = T)
}
\arguments{
\item{open}{If open the dotto file in Rstudio or not. Default is \code{TRUE}.}
}
\description{
It creates a template of \code{Dotto} based on an existing \code{Dotto} and we call them a family.
The new \code{Dotto} automatically has a line connection to the base \code{Dotto} and the card generated in
the index page combine the information related to all family.
}
