% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scc_hexsticker.R
\name{create_scc_sticker}
\alias{create_scc_sticker}
\title{Creating an scc style sticker for R packages}
\usage{
create_scc_sticker(imgurl, package_name)
}
\arguments{
\item{imgurl}{url of the scc logo}

\item{package_name}{name of the package}
}
\description{
Creating an scc style sticker for R packages
}
