% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch00dataFormat.R
\name{dataFormat}
\alias{dataFormat}
\title{dataFormat}
\usage{
dataFormat(data, na = NULL, id = 1, Z = NULL, w = NULL)
}
\arguments{
\item{data}{is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{id}{id indicates the column number containing the examinee ID. The default is 1.
If the answer pattern is contained in the first column, it is treated as if there is no ID vector.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
\describe{
\item{U}{Data matrix. A matrix with rows representing the sample size and columns
representing the number of items, where elements are either 0 or 1. \eqn{u_{ij}=1} indicates
that student i correctly answered item j, while \eqn{u_{ij}=0} means that student i answered
item j incorrectly. If the data contains NA values, any value can be filled in the matrix U,
represented by the following missing value index matrix Z. However, in this funciton,
-1 is assigned.}
\item{ID}{The ID label given by the designated column or function."}
\item{ItemLabel}{The item names given by the provided column names or function.}
\item{Z}{Missing indicator matrix.\eqn{z_{uj}=1} indicates that item j is presented to Student i,
while \eqn{z_{ij}=0} indicates item j is NOT presented to Student i.}
\item{w}{item weight vector}
}
}
\description{
This function serves the role of formatting the data prior to the analysis.
}
