% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_DataMotto.R
\docType{package}
\name{DataMotto}
\alias{DataMotto}
\title{DataMotto: A package for generating Data Motto website}
\description{
The DataMotto package provides two main template \code{\link{Index}}
and \code{\link{Dotto}} for creating the index and post pages with a
custom css/js. It also provides many utility functions that help creating the website
more interactive.
}
