% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDesc.R
\docType{data}
\name{DatasetDesc}
\alias{DatasetDesc}
\title{Dataset template description}
\format{An object of class \code{data.frame} with 13 rows and 4 columns.}
\usage{
data(DatasetDesc)
}
\description{
Description of all the fields available in the template for data input.
This table is for information purposes only.
}
\examples{
data(DatasetDesc)
}
\keyword{dataset}
\keyword{description}
\keyword{fields}
