% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{default_discrete_palette}
\alias{default_discrete_palette}
\title{Color palette for discrete features}
\format{An object of class \code{character} of length 45.}
\usage{
default_discrete_palette
}
\description{
Color palette for discrete features
}
\keyword{datasets}
