% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_template.R
\name{default_template}
\alias{default_template}
\title{Shared HTML template function}
\usage{
default_template(
  template_name,
  template_path,
  template_dependencies,
  pandoc_args,
  in_header,
  before_body,
  after_body,
  ...
)
}
\description{
Shared HTML template function
}
