% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{DefineModel}
\alias{DefineModel}
\title{A Function to define the population-based threshold model(PBTM) you want to work with.}
\usage{
DefineModel(ModelCode)
}
\arguments{
\item{DefineModel(1.1)}{selects the Thermal time Sub-optimal Model}

\item{DefineModel(2.1)}{selects the Suboptimal Hydrotime Model}
}
\value{
Sets up the package and print the selected model.
}
\description{
This function allows you to choose the PBT model to be calculated. Choose the model that you want to work on (6.1)Hydropriming model; (2.1)Suboptimal Hydrotime; (2.2)Supra-optimal Hydrotime; (1.1)Thermaltime; (3.1)Suboptimal Hydrothermal Time; (3.2)Supra-optimal Hydrothermal Time.
}
\examples{
DefineModel(2.1)
}
\keyword{PBTM}
