% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dicho}
\alias{dicho}
\title{Internal function - estimation of the parameter a}
\usage{
dicho(x, y, threshold, nb, left_bound, right_bound)
}
\arguments{
\item{x}{a matrix of observations. Columns correspond to biological samples and rows to microorganisms.}

\item{y}{y.}

\item{threshold}{threshold.}

\item{nb}{nb.}

\item{left_bound}{left_bound.}

\item{right_bound}{right_bound.}
}
\value{
a numeric.
}
\description{
Internal function - estimation of the parameter a
}
\keyword{internal}
