% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precip.R
\name{do_precip}
\alias{do_precip}
\title{internal Cyclones function, called by get_precip to route function calls for different data sources and to format final product}
\usage{
do_precip(
  cent,
  source,
  r,
  todir = NULL,
  dpath = NULL,
  overwrite = FALSE,
  loadrsts = FALSE,
  pre_storm
)
}
\description{
'do_precip()' is currently only called by the 'get_precip()' function and its parallel helper 'parallelprecip()'. The function routes processing to the
appropriate data acquisition function depending on the desired data source, then applies the same set of routines to ensure all products are formatted consistently.
}
