% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_2015pop_old.R
\name{download_2015pop_old}
\alias{download_2015pop_old}
\title{Download 2015 Population Data per Each Old Municipality}
\usage{
download_2015pop_old(pref_code)
}
\arguments{
\item{pref_code}{administrative code for prefecture of choice (eg. Hokkaido: 01, Okinawa: 47)}
}
\value{
shapefile csv downloaded in data/ directory and resulting csv
}
\description{
Download 2015 Population Data per Each Old Municipality
}
\concept{downloaddata}
