% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_2020_census.R
\name{download_2020_census}
\alias{download_2020_census}
\title{Download Prefectural Population Demographics Data}
\usage{
download_2020_census(type)
}
\arguments{
\item{type}{character vector for the type of data either 'total' or 'foreigner'.}
}
\value{
data frame with the 2020 census excel file.
}
\description{
Download Prefectural Population Demographics Data
}
\concept{downloaddata}
