% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_old_shp.R
\name{download_old_shp}
\alias{download_old_shp}
\title{Download 2000 Prefectural Shapefile Data}
\usage{
download_old_shp(pref_code)
}
\arguments{
\item{pref_code}{administrative code for prefecture of choice (eg. Hokkaido: 01, Okinawa: 47)}
}
\value{
shapefile data downloaded in data/ directory and resulting shp
}
\description{
Download 2000 Prefectural Shapefile Data
}
\concept{downloaddata}
