% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{draw.ROC_SINBA}
\alias{draw.ROC_SINBA}
\title{draw.ROC_SINBA}
\usage{
draw.ROC_SINBA(
  ac_mat_sinba = NULL,
  ac_mat_single = NULL,
  SINBA_originalID = NULL,
  phenotype_info = NULL,
  group_col = NULL,
  G1_name = NULL,
  G0_name = NULL,
  main = "",
  pdf_file = NULL
)
}
\arguments{
\item{ac_mat_sinba, }{numeric matrix, the activity values of the selected seed-partner driver across all samples.}

\item{ac_mat_single, }{numeric matrix, the activity values of single driver across all samples}

\item{SINBA_originalID, }{seed and parterne driver combination, is generated by \code{combineNet2target_list} with return_type="SINBA_originalID"}

\item{phenotype_info, }{a data frame with sample group information}

\item{group_col, }{character, the column name to be used for group}

\item{G1_name, }{a vector of characters, the sample names of experimental group.}

\item{G0_name, }{a vector of characters, the sample names of NON-experimental group.}

\item{main, }{character, the main title of the plot}

\item{pdf_file, }{character, the file path to save as PDF file. If NULL, no PDF file will be save. Default is NULL.}
}
\description{
draw.ROC_SINBA draws ROC curves, including both seed and partner drive curves and the combination driver curve.
}
\examples{
ac_mat_sinba<-exprs(SINBA.par$SINBA.ac.eset)
ac_mat_single<-exprs(SINBA.par$SINBA_single.ac.eset)
SINBA_originalID<-"LCK:NOTCH1"
phenotype_info<-pData(SINBA.par$SINBA.ac.eset)
phenotype_info$condition<-ifelse(phenotype_info$das_lc50>80,"Resistant","Sensitive")
draw.ROC_SINBA(ac_mat_sinba=ac_mat_sinba,ac_mat_single=ac_mat_single,SINBA_originalID=SINBA_originalID,phenotype_info=phenotype_info,group_col="condition",G1_name="Sensitive",G0_name="Resistant",pdf_file=sprintf("./ROC_\%s.pdf",SINBA_originalID))
}
