\name{@EIP_NAME_SHORT@-package}
\alias{@EIP_NAME_SHORT@-package}
\alias{@EIP_NAME_SHORT@}
\docType{package}
\title{@EIP_NAME@}
\description{@EIP_DESCRIPTION@}
\details{
\tabular{ll}{
Package: \tab @EIP_NAME_SHORT@\cr
Type: \tab Package\cr
Version: \tab @EIP_VER_MAJOR@.@EIP_VER_MINOR@.@EIP_VER_DATE@.@EIP_VER_GIT@\cr
Date: \tab @EIP_VER_DATE@\cr
License: \tab @EIP_LICENSE@\cr
}
Joel's Overview
}
\author{
@EIP_AUTHOR@
}
\references{
Some references
}
Ecology, phenology
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{}
