\name{eip.version}
\alias{eip.version}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Package version }
\description{ Return the pacage version }
\usage{ eip.version() }
%- maybe also 'usage' for other objects documented here.
\arguments{}
\value{ Returns the package version}
%%\references{}
\author{@EIP_AUTHOR@}
%%\note{}

\seealso{
\code{\link{eip.showModel}}
\code{\link{eip.genOutput}}
\code{\link{eip.histcmp}}
\code{\link{eip.plot}}
\code{\link{eip.smooth}}
}
\examples{}
\keyword{ EcoIP }
