% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_2020_pop.R
\name{estimate_2020_pop}
\alias{estimate_2020_pop}
\title{Estimate 2020 Populations For Lower-Level Census Units from Municipality-Level Data}
\usage{
estimate_2020_pop(pref, census2020)
}
\arguments{
\item{pref}{an sf object with cleaned, collated version of the census data}

\item{census2020}{the output of clean_2020_census}
}
\value{
vector of population estimates
}
\description{
Estimate 2020 Populations For Lower-Level Census Units from Municipality-Level Data
}
\concept{estimatedata}
