% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_message_pull.R
\name{facebook_message_pull}
\alias{facebook_message_pull}
\title{facebook_message_pull}
\usage{
facebook_message_pull(folder)
}
\arguments{
\item{folder}{filepath to data}
}
\value{
facebook message conversation to a .csv file
}
\description{
Facebook: Message Conversation Extractor
}
\details{
This function pulls the sentences from your messages folder and places them in a handy .csv file for analysis.
The participants, timestamp, and content of the messages are pulled out and placed in the .csv file that is created.
The .csv file is names for the participant you have been messaging. The function also pulls the date and time from the Json file
and converts it and writes the Sender, Time, and Content of the message to an .xlsx file.
}
\examples{

library(facebookanalysis)

\dontrun{facebook_message_pull(folder = data)}

}
\keyword{facebook_message_pull}
