% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facebook_sentences_pull.R
\name{facebook_sentences_pull}
\alias{facebook_sentences_pull}
\title{facebook_sentences_pull}
\usage{
facebook_sentences_pull(folder)
}
\arguments{
\item{folder}{filepath to data}
}
\value{
facebook sentences
}
\description{
Facebook: Sentiment Calculator
}
\details{
This function pulls the sentences from your messages folder created by the facebook_message_pull function.
The sentences are then placed in a csv file and written to the nrc_sentences folder that is created by the function.
}
\examples{

library(facebookanalysis)

\dontrun{facebook_sentences_pull(folder = data)}

}
\keyword{facebook_sentences_pull}
