% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{FixedTb}
\alias{FixedTb}
\title{A Function to manually define the temperature base (Tb) for model calculation.}
\usage{
FixedTb(Tb1)
}
\arguments{
\item{FixedTb(11)}{will fix the Tb to 11C for further calculation of the model.}
}
\description{
This function defines the Tb value and avoid it to be calculated together with other values.
}
\examples{
FixedTb(11)
FixedTb()
}
\keyword{Known}
\keyword{Tb}
\keyword{base}
\keyword{fixed}
\keyword{temperature}
