% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{foo_a}
\alias{foo_a}
\title{Useful function to estimate the parameter a}
\usage{
foo_a(x, nb, left_bound, right_bound)
}
\arguments{
\item{x}{x.}

\item{nb}{nb.}

\item{left_bound}{left_bound.}

\item{right_bound}{right_bound.}
}
\value{
a numeric.
}
\description{
Useful function to estimate the parameter a
}
\keyword{internal}
