% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{geneCategories}
\alias{geneCategories}
\title{geneCategories}
\usage{
geneCategories()
}
\description{
the available geneCategories used in DGIdb. Users can select specific geneCategories
}
\examples{
geneCategories()
}
