% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_font.R
\name{get_scc_font}
\alias{get_scc_font}
\title{Load the scc's font Source Sans Pro}
\usage{
get_scc_font()
}
\description{
Load the scc's font Source Sans Pro
}
\details{
function that obtains the font files in inst/fonts
Source Sans Pro is the font currently used on the SCC website. This
function gets called in \code{scc_style()}
}
