% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{IIF2PLM}
\alias{IIF2PLM}
\title{IIF for 2PLM}
\usage{
IIF2PLM(a, b, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{locaiton parameter}

\item{theta}{ability parameter}
}
\description{
Item Information Function for 2PLM
}
