% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobiclustering.R
\name{is.cobiclustering}
\alias{is.cobiclustering}
\title{Is an object of class cobiclustering ?}
\usage{
is.cobiclustering(object)
}
\arguments{
\item{object}{an object of class cobiclustering.}
}
\description{
Is an object of class cobiclustering ?
}
\keyword{internal}
