% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMclus.R
\name{itemEll}
\alias{itemEll}
\title{calc final item-ell}
\usage{
itemEll(U, Z, postDist, classRefMat)
}
\arguments{
\item{U}{response matrix U of the examData class.}

\item{Z}{missing indicator matrix Z of the examData class.}

\item{postDist}{class membership matrix}

\item{classRefMat}{class reference matrix}
}
\description{
Using the original data, class membership matrix, and class reference matrix,
the log-likelihood for each item is calculated.s
}
