% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_explain.R
\docType{data}
\name{J5S10}
\alias{J5S10}
\title{J5S10.Rdata}
\format{
A data frame with 5 students and 10 items
}
\source{
\url{http://shojima.starfree.jp/tde/index.htm}
}
\usage{
J5S10
}
\description{
J5S10.Rdata
}
\keyword{datasets}
