% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04AbilityEstimation.R
\name{LLtheta_mat}
\alias{LLtheta_mat}
\title{log-likelihood function for ability parameter}
\usage{
LLtheta_mat(theta, Lambda, U, Z)
}
\arguments{
\item{theta}{target unknown parameter}

\item{Lambda}{Item parameters matrix}

\item{U}{data matrix}

\item{Z}{missing indicator}
}
\description{
An internal function to calculate the log-likelihood
}
