% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\docType{data}
\name{Models}
\alias{Models}
\title{Models Description}
\format{An object of class \code{data.frame} with 11 rows and 4 columns.}
\usage{
data(Models)
}
\description{
Description of all models, respective codes for the package use and references.
}
\examples{
data(Models)
}
\references{
Bello and Bradford (2016) Seed Science Research 26(3) 199–221.
(\href{http://dx.doi.org/10.1017/ S0960258516000179.}{Cambridge})
}
\keyword{PBTM}
\keyword{models}
