% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\docType{data}
\name{morphology_obs}
\alias{morphology_obs}
\title{Morphology data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 72 rows and 12 columns.
}
\usage{
morphology_obs
}
\description{
Morphology data
}
\keyword{datasets}
