% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\docType{data}
\name{motility_obs}
\alias{motility_obs}
\title{Motility data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 85 rows and 12 columns.
}
\usage{
motility_obs
}
\description{
Motility data
}
\keyword{datasets}
