% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_cut_edge.R
\name{original_cut_edge}
\alias{original_cut_edge}
\title{Calculate Original Compactness (cut edge)}
\usage{
original_cut_edge(pref, pop_tol)
}
\arguments{
\item{pref}{an sf object with cleaned, collated version of the census data}

\item{pop_tol}{the vector of precinct populations. Defaults to the pop, population, or total_pop columns, if one exists.}
}
\value{
a vector with the compactness score for the vector of plans
}
\description{
Calculate Original Compactness (cut edge)
}
\concept{calculate}
