% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_weight_disparity.R
\name{original_weight_disparity}
\alias{original_weight_disparity}
\title{Calculate Original Ippyo-no-kakusa (maximum weight disparity in voter power)}
\usage{
original_weight_disparity(pref, parameter)
}
\arguments{
\item{pref}{an sf object with cleaned, collated version of the census data}

\item{parameter}{character vector of parameter (\code{max_to_min}, \code{Gini}, \code{LH}, and \code{HH}).}
}
\value{
a vector with the ippyo no kakusa for the original plan
}
\description{
Calculate Original Ippyo-no-kakusa (maximum weight disparity in voter power)
}
\concept{calculate}
