% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{palette_OkabeIto}
\alias{palette_OkabeIto}
\title{Color palette from colorblindr}
\format{An object of class \code{character} of length 8.}
\usage{
palette_OkabeIto
}
\description{
Color palette from colorblindr
}
\keyword{datasets}
