% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02TestItemAnalysis.R
\name{passage}
\alias{passage}
\title{Passage Rate of student}
\usage{
passage(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\value{
This function returns passage rate for each students
}
\description{
Passage rate of Student s is NRS divided by the number of presented items.
}
\examples{
# passage(U)
}
