% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{penalty}
\alias{penalty}
\title{Calculate the BIC penalty}
\usage{
penalty(x)
}
\arguments{
\item{x}{an object of class biclustering.}
}
\value{
the value of the BIC penalty.
}
\description{
Calculate the BIC penalty
}
\keyword{internal}
