% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotActualvsModelPredicted}
\alias{PlotActualvsModelPredicted}
\title{A Function to plot actual and model predicted germination.}
\usage{
PlotActualvsModelPredicted()
}
\description{
This function plots germination time points predicted by the model with actual germination percentage.
}
\examples{
PlotActualvsModelPredicted()
PlotActualvsModelPredicted()
}
\keyword{actual}
\keyword{data}
\keyword{model}
\keyword{plot}
\keyword{predicted}
