% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotActualvsModelPredictedTreatments}
\alias{PlotActualvsModelPredictedTreatments}
\title{A Function to plot actual and model predicted germination.}
\usage{
PlotActualvsModelPredictedTreatments()
}
\description{
This function plots germination time points predicted by the model with actual germination percentage with treatments.
}
\examples{
PlotActualvsModelPredictedTreatments()
PlotActualvsModelPredictedTreatments()
}
\keyword{actual}
\keyword{data}
\keyword{model}
\keyword{plot}
\keyword{predicted}
