% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotCleanData}
\alias{PlotCleanData}
\title{A Function to plot the cleaned data selected.}
\usage{
PlotCleanData()
}
\arguments{
\item{PlotCleanData()}{will plot cleaned cumulative curves over time.}
}
\description{
This function plots the cleaned data previously selected after removing repetitive percentages. Only intial presence of the percentage is displayed.
}
\examples{
PlotCleanData()
PlotCleanData()
}
\keyword{cleaned}
\keyword{data}
\keyword{plot}
