% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotCleanModel}
\alias{PlotCleanModel}
\title{A Function to plot the selected calculated model and parameters and predicitions with the cleaned data.}
\usage{
PlotCleanModel()
}
\description{
This function plots the selected model and calculated parameters with the cleaned data.
}
\examples{
PlotCleanModel()
PlotCleanModel()
}
\keyword{model}
\keyword{plot}
\keyword{population-based}
\keyword{threshold}
