% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotGR50vsTemp}
\alias{PlotGR50vsTemp}
\title{A Function to plot GR50 vs temperature.}
\usage{
PlotGR50vsTemp()
}
\description{
This function plots germination rates against temperature.
}
\examples{
PlotGR50vsTemp()
PlotGR50vsTemp()
}
\keyword{GR50}
\keyword{Temperature}
\keyword{plot}
