% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotNormalizedTime}
\alias{PlotNormalizedTime}
\title{A Function to plot model normalized time with actual germination data.}
\usage{
PlotNormalizedTime()
}
\description{
This function plots normalized time predicted by the model with actual germination percentage.
}
\examples{
PlotNormalizedTime()
PlotNormalizedTime()
}
\keyword{actual}
\keyword{data}
\keyword{model}
\keyword{normalized}
\keyword{plot}
\keyword{predicted}
\keyword{time}
