% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotPBTMModel}
\alias{PlotPBTMModel}
\title{A Function to plot the selected calculated model and parameters and predicitions.}
\usage{
PlotPBTMModel()
}
\description{
This function plots the selected model and calculated parameters.
}
\examples{
PlotPBTMModel()
PlotPBTMModel()
}
\keyword{model}
\keyword{plot}
\keyword{population-based}
\keyword{threshold}
