% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotPrimingMatrix}
\alias{PlotPrimingMatrix}
\title{A Function to plots the hydropriming and hydrothermal priming models in a 3D surface graph.}
\usage{
PlotPrimingMatrix()
}
\description{
This function plots the germination rate (GR50) for each temperature (when more than one temperature was used) in relation to priming duration and water potential.
}
\examples{
PlotPrimingMatrix()
PlotPrimingMatrix()

}
\keyword{3D}
\keyword{hydropriming}
\keyword{hydrothermal}
\keyword{models}
\keyword{priming}
\keyword{surface}
