% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBTM-functions.R
\name{PlotPrimingModel}
\alias{PlotPrimingModel}
\title{A Function to plot the both priming models.}
\usage{
PlotPrimingModel()
}
\description{
This function plots the priming models and calculated parameters.
}
\examples{
PlotPrimingModel()
PlotPrimingModel()
}
\keyword{hydropriming}
\keyword{hydrothermal}
\keyword{model}
\keyword{plot}
\keyword{priming}
