% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrimingDt.R
\docType{data}
\name{PrimingDt}
\alias{PrimingDt}
\title{Priming data provided by a collaborator on different combinations of temperature, water potential and durations}
\format{An object of class \code{data.frame} with 259 rows and 13 columns.}
\usage{
data(PrimingDt)
}
\description{
Data from germination experiments done at 20C and pure water.
}
\examples{
data(PrimingDt)
}
\references{

}
\keyword{Priming}
\keyword{data}
\keyword{datasets}
\keyword{germination}
\keyword{rate}
