% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwbGS.R
\name{pwbGS}
\alias{pwbGS}
\title{pwbGS}
\usage{
pwbGS(theta, des, bounds, nsims = 1e+05)
}
\arguments{
\item{theta}{true response probablity}

\item{des}{a realisation of a GS design, formatted as per output of curtailment::singlearmDesign}

\item{bounds}{a data frame with 3 cols: m, fail, success, corresponding to
cumulative sample size at each interim analysis, and fail/success boundaries}

\item{nsims}{number of simulations (default 1e5)}
}
\value{
list of length 3:
results: individual sims
ests: estimates
mc.error
}
\description{
For a single group sequential design, obtains bias for a single true response rate.
}
