% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{queryDgidb_V5_drugAnnotation}
\alias{queryDgidb_V5_drugAnnotation}
\title{queryDgidb_V5_drugAnnotation}
\usage{
queryDgidb_V5_drugAnnotation(drugs)
}
\arguments{
\item{drugs}{character vector, the query drug names.}
}
\description{
Using ghql GraphqlClient function to query DGIdb version 5 to find the drug meta information, including: "name","conceptId","approved" ,"immunotherapy", "antiNeoplastic","drugAttributes" ,"drugApprovalRatings" ,"drugApplications".
}
\examples{
use_drugs<-c("Dasatinib","DOVITINIB","nondrug")
res<-queryDgidb_V5_drugAnnotation(drugs=use_drugs)

}
