% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPW.R
\name{randomiseRPW}
\alias{randomiseRPW}
\title{randomiseRPW}
\usage{
randomiseRPW(N, Pa, Pb)
}
\arguments{
\item{N}{sample size}

\item{Pa}{response probabaility for arm A}

\item{Pb}{response probabaility for arm B}
}
\value{
Vector of results: number of patients and responses in arm A and
number of patients and responses in arm B.
}
\description{
Simulates a randomised play-the-winner trial
}
