% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_hpp_dat.R
\name{raw_hpp_dat_date}
\alias{raw_hpp_dat_date}
\title{Check what date the latest file is}
\usage{
raw_hpp_dat_date(path)
}
\arguments{
\item{path}{From \link{raw_hpp_dat_path}.}
}
\description{
Check what date the latest file is
}
