% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflect_old_boundaries.R
\name{reflect_old_boundaries}
\alias{reflect_old_boundaries}
\title{Download 2015 Population Data per Each Old Municipality}
\usage{
reflect_old_boundaries(
  pref,
  old_boundary,
  pop_by_old_boundary,
  old_code,
  new_code
)
}
\arguments{
\item{pref}{sf object of cleaned census data for a prefecture}

\item{old_boundary}{sf object of administrative boundaries as of the year 2000}

\item{pop_by_old_boundary}{sf object of population per each former-city}

\item{old_code}{the codes of former-cities that were later merged together}

\item{new_code}{the code of the new city that was created as a result of gappei}
}
\value{
shapefile csv downloaded in data/ directory and resulting csv
}
\description{
Download 2015 Population Data per Each Old Municipality
}
\concept{downloaddata}
