% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_posts.R
\name{render_posts}
\alias{render_posts}
\title{Render Gallery Posts}
\usage{
render_posts(time = 0, eval_all = FALSE)
}
\arguments{
\item{time}{(\code{integer(1)})\cr
A posts is skipped when the file was rendered less than \code{time} minutes ago.}

\item{eval_all}{(\code{logical(1)})\cr
If \code{TRUE}, all chunks are evaluated and intermediate results are stored.}
}
\description{
Renders all rmarkdown files in the \verb{_posts}  directory.
}
