% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_piechart.R
\name{scc_piechart}
\alias{scc_piechart}
\title{Define piechart in scc colours}
\usage{
scc_piechart(
  df,
  value,
  group,
  title = NULL,
  subtitle = NULL,
  format = "report"
)
}
\arguments{
\item{df}{The dataframe provided}

\item{value}{The value of each group, no need to convert into a percentage}

\item{group}{The associated group for each value}

\item{title}{The title that should go above the plot}

\item{subtitle}{The subtitle that should go below the title}

\item{format}{can be one of "report", "presentation" or "tweet". Will adjust
font sizes approriately for that format.}
}
\description{
Define piechart in scc colours
}
