% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{selectSP_driver_byDB}
\alias{selectSP_driver_byDB}
\title{selectSP_driver_byDB}
\usage{
selectSP_driver_byDB(
  genes = NULL,
  select_type = NULL,
  database = NULL,
  select_phase = NULL
)
}
\arguments{
\item{genes}{a character vector of genes for which the targeting drug phase or blood brain barrier penetration information are queried.}

\item{select_type}{character, which drug information is required. User can choose from "BP_prediction" and "drug_phase".}

\item{database}{character, which database will be used. "BPdb" contains drug blood brain barrier penetrantion information, others contain drug developing phase information. User can choose from "BPdb", "ChEMBL","Repurposing_HUB" and "Drugbank_V5.1".}

\item{select_phase}{a character vector of drug phases will be used. For checking detailed phase information in each database, try to call helper functions: \code{ChEMBL_drugphase_levels()}; \code{Repurposing_HUB_drugphase_levels()}; \code{Drugbank_drugphase_levels()}.}
}
\description{
selectSP_driver_byDB is used to select driver genes based on their interaction drug information, including predicted blood-brain barrier penetration and drug developing phase.
}
\examples{
drugdb.preload()
genes<-c("MDM2","CDK4","CDK6","ATM","ATR","BCL2","BCL2L1","HDAC6","STAT3","JAK2")
sp_df<-selectSP_driver_byDB(genes=genes,select_type="BP_prediction",database="BPdb")
sp_df<-selectSP_driver_byDB(genes=genes,select_type="drug_phase",database="Drugbank_V5.1")
sp_df<-selectSP_driver_byDB(genes=genes,select_type="drug_phase",database="Drugbank_V5.1",select_phase="approved")
sp_df<-selectSP_driver_byDB(genes=genes,select_type="drug_phase",database="ChEMBL")
sp_df<-selectSP_driver_byDB(genes=genes,select_type="drug_phase",database="Repurposing_HUB")
\dontrun{
}
}
