% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progreesbar.R
\name{show.progress}
\alias{show.progress}
\title{progress-bar function}
\usage{
show.progress(pos, len, msg)
}
\arguments{
\item{pos}{position}

\item{len}{total length}

\item{msg}{show message}
}
\description{
A progress bar to display the calculation process.
}
