% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04ParameterEstimation.R
\name{slopeprior}
\alias{slopeprior}
\title{Prior distribution function with respect to the slope.}
\usage{
slopeprior(a, m, s, const = 1e-15)
}
\arguments{
\item{a}{slope coefficient}

\item{m}{prior parameter to be set}

\item{s}{prior parameter to be set}

\item{const}{A very small constant}
}
\description{
Prior distribution function with respect to the slope.
}
