% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SINBA.functions_0.1.4.R
\name{sourceDatabases}
\alias{sourceDatabases}
\title{sourceDatabases}
\usage{
sourceDatabases()
}
\description{
the available source database used in DGIdb Users can select partial of the source database
}
\examples{
sourceDatabases()
}
