% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch02TestItemAnalysis.R
\name{StudentAnalysis}
\alias{StudentAnalysis}
\title{StudentAnalysis}
\usage{
StudentAnalysis(U, na = NULL, Z = NULL, w = NULL)
}
\arguments{
\item{U}{U is a data matrix of the type matrix or data.frame.}

\item{na}{na argument specifies the numbers or characters to be treated as missing values.}

\item{Z}{Z is a missing indicator matrix of the type matrix or data.frame}

\item{w}{w is item weight vector}
}
\description{
The StudentAnalysis function returns descriptive statistics for each individual student.
Specifically, it provides the number of responses, the number of correct answers,
the passage rate, the standardized score, the percentile, and the stanine.
}
