% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syllabus.R
\name{syllabus}
\alias{syllabus}
\title{Load class syllabus URL in browser}
\usage{
syllabus()
}
\value{
opens URL of class syllabus in browser
}
\description{
Load class syllabus URL in browser
}
\examples{
syllabus()
}
