% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_overlap.R
\name{temp_overlap}
\alias{temp_overlap}
\title{Assemblage-wide temporal niche segregation}
\usage{
temp_overlap(mat, method = c("pianka", "czekanowski"))
}
\arguments{
\item{mat}{A matrix with species per time columns.}

\item{method}{A measure of temporal niche overlap between pairs of species}
}
\value{
A matrix with pairwise overlaping temporal resource use.
}
\description{
Assemblage-wide temporal niche segregation
}
\examples{
n = 200
k = 200
m = matrix(ifelse(runif(n*k) > 0.5, 1, 0), ncol = k)
temp_overlap(m, method = "pianka")
temp_overlap(m, method = "czekanowski")
}
