% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{TestInformationFunc}
\alias{TestInformationFunc}
\title{TIF for IRT}
\usage{
TestInformationFunc(params, theta)
}
\arguments{
\item{params}{parameter matrix}

\item{theta}{ability parameter}
}
\description{
Test Information Function for 4PLM
}
