% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ch04IRTbaseFunctions.R
\name{ThreePLM}
\alias{ThreePLM}
\title{Three-Parameter Logistic Model}
\usage{
ThreePLM(a, b, c, theta)
}
\arguments{
\item{a}{slope parameter}

\item{b}{locaiton parameter}

\item{c}{lower asymptote parameter}

\item{theta}{ability parameter}
}
\description{
The three-parameter logistic model is a model where the lower
asymptote parameter c is added to the 2PLM
}
