% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TomatoQ2Dt.R
\docType{data}
\name{TomatoQ2Dt}
\alias{TomatoQ2Dt}
\title{Tomato seed respiration data on different combinations of temeporature and water potential}
\format{An object of class \code{data.frame} with 418 rows and 13 columns.}
\usage{
data(TomatoQ2Dt)
}
\description{
Data from Q2 experiments on water and temperature stress in
Tomato, with data on 3 temperatures (15, 20 and 25C) and 3 water potentials (0, -0.25, 0.4).
The outcome is the fraction of seeds reaching 50% oxygen consumption after imbibition over time.
}
\examples{
data(TomatoQ2Dt)
}
\references{
Bello and Bradford (2016) Seed Science Research 26(3) 199–221.
(\href{http://dx.doi.org/10.1017/ S0960258516000179.}{Cambridge})
}
\keyword{Q2}
\keyword{Seed}
\keyword{Tomato}
\keyword{datasets}
\keyword{respiration}
