% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_Dotto.R
\name{update_Dotto}
\alias{update_Dotto}
\title{Update Dotto file and folder names based on updated \code{slug}}
\usage{
update_Dotto(path, clean_dotto_deps = T)
}
\arguments{
\item{path}{a path to the Dotto \code{.Rmd} file.}

\item{clean_Dotto_deps}{Delete all the files and folders in Dotto and only keep .Rmd file. Default is \code{FALSE}.}
}
\description{
Update Dotto file and folder names based on updated \code{slug}
}
