% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dotto.R
\name{use_dotto}
\alias{use_dotto}
\title{Dotto extensions}
\usage{
use_dotto(metadata)
}
\description{
Adds required extensions that are necessary for setting up
Rmd file gets \code{Dotto} style.
}
\details{
Add the following code chunk to your Rmarkdown file with output \code{DataMotto::Dotto}.
\code{```{r datamotto_dotto, echo=FALSE}
        DataMotto::use_dotto()
        ```}
}
\examples{
\dontrun{
DataMotto::use_dotto()
}
}
