% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\docType{data}
\name{volume_obs}
\alias{volume_obs}
\title{Volume data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43 rows and 12 columns.
}
\usage{
volume_obs
}
\description{
Volume data
}
\keyword{datasets}
