% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weirds.R
\name{weird}
\alias{weird}
\title{Define a call to a weird function}
\usage{
weird(method, ...)
}
\arguments{
\item{method}{- weird method to be used, check list of available
methods by using \code{\link{weirds_list}}}

\item{\dots}{- additional parameters to be passed to weird method}
}
\value{
list with class weirdSpecs to be used as value when using
the parameter \code{tuneList} of \code{\link{stranger}} function.

For every \code{weird} function, possible parameters in \dots should be checked by looking to the help of the underlying function. Some special common parameters are \code{info} to be set to TRUE to have information about the method, inclusing the name and package of the main underlying function. Another common parameter  is  \code{colname} that allow to overide default prefix name used for the outcome of the computation.

Methods specific parameters, additional parameters may be used (for instance \code{simplify} for _knn_ method or \code{type} for kmeans. See the vignette describing the list of available _weirds_ methods in base package.
}
\description{
weirds function are wrapper around other packages anomalies
detection routines. \code{weird} function is intended to be
used when invoking \code{\link{stranger}} so that several anomalies
detection routines are used at one time.
}
