% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_config_posts.R
\name{config_posts}
\alias{config_posts}
\title{Config all posts (Dottos)}
\usage{
config_posts(title_width = 50, desc_width = 150)
}
\description{
It creates a \code{posts.json} file in the root the package with
all the high level information (\code{metadata}) of posts.

@param title_width numeric. The maximum width of title shows in the card of index page. Default is \code{50} characters.
@param desc_width numeric. The maximum width of description shows in the card of index page. Default is \code{150} characters.
@importFrom here here
@importFrom jsonlite toJSON
@import rmarkdown
@importFrom glue glue
}
