% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cons_stormdat.R
\name{cons_stormdat}
\alias{cons_stormdat}
\title{Consolidate storm tabular data.}
\usage{
cons_stormdat(
  dat,
  vars = c("wind", "pres", "rmw", "quads", "roci", "poci", "eye"),
  msw_int = "1min",
  pref = c("USA", "WMO"),
  fun = "mean"
)
}
\arguments{
\item{dat}{A singular dataframe or listed data set, sourced from IBTrACS and containing multiple analogous values from different agencies.}

\item{vars}{Short names for the variables to be consolidated. Values of 'wind', 'pres', and 'rmw' correspond to the maximum sustained
wind, the minimum central pressure, and the radius of maximum wind, respectively. A value of 'quads' refers to the quadrant specific
wind extents of 34, 50, and 64 knot winds in the NE, SE, SW, and NW quadrants.}

\item{msw_int}{The target interval for the maximum sustained winds, which is not consistent among the agencies. The USA uses a one-minute
interval, most other agencies use a 10-minute interval, but 2-minute and 3-minute intervals are also used. When consolidating the wind columns,
linear models are used to translate any values whose interval is not the target interval. The models are computed from all rows in which multiple
interval values are provided.}

\item{pref}{Agency preferences for all values, listed in order of preference. If these values are present, they will be used (after conversion to the
target interval). These can be any of the agency or model abbreviations in the IBTrACS data (USA, WMO, TOKYO, CMA, HKO, KMA, NEWDELHI, REUNION, BOM,
NADI, WELLINGTON, DS824, TD9636, TD9635, NEUMANN, MLC). However, because storms sometimes traverse multiple agency jurisdictions, WMOfirst or WMOlast
can also be used to avoid sudden changes in values.}

\item{fun}{The summary function to be applied to any instances of multiple values.}
}
\value{
A data frame or listed data set, whatever the input was, with new columns whose names are prefixed with 'CONS_' representing the consolidated values.
}
\description{
\code{cons_stormdat()} will accept a singular data frame sourced from IBTrACS or a listed data set produced by 'get_storms()' and will consolidate
the desired variables across the various agencies into one column. In the case of multiple values for a given variable, either the preferred agency
is used or a function is applied to summarise the values.
}
