% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_create_Dotto.R
\name{create_Dotto}
\alias{create_Dotto}
\title{Create a new Dotto}
\usage{
create_Dotto(title = NULL, lang = "r", open = T)
}
\arguments{
\item{title}{Title of the dotto post.}

\item{lang}{The programming language for writing the Dotto. Default is \code{R}. Other
options are \code{python}, \code{julia}, \code{sql}, \code{bash}, \code{js}, \code{node}, \code{d3}, \code{Rcpp}, \code{stan}. It is not
case-sensitive.}

\item{open}{If open the dotto file in Rstudio or not. Default is \code{TRUE}.}
}
\description{
It creates a dotto (.Rmd) based on a default \code{slug}, created automatically based on title.
}
\details{
In the \code{yml} section, here is the avaialbe values for some tags:
\itemize{
\item \code{lang}: \code{R}, \code{python}, \code{julia}, \code{sql}, \code{bash}, \code{js}, \code{node}, \code{d3}, \code{Rcpp}, \code{stan}
\item \code{categories}: \code{Importing-Cleaning}, \code{Manipulation}, \code{Visualization}, \code{Machine-Learning}, \code{Statistical-Learning}, \code{Reporting}, \code{Data-Engineering}
\item \code{url}: A personal url address that can be for: your persoanl website, twitter account, linkedin, etc.
\item \code{applications}: Having an specific applications on files \code{Finance}, \code{Pharma}.
}
}
