% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_create_family.R
\name{create_family}
\alias{create_family}
\title{Create a \code{Dotto} family be generating a new \code{Dotto} based on another existing \code{Dotto}.}
\usage{
create_family(Dotto_base_label, open = T)
}
\arguments{
\item{Dotto_base_label}{The label (or path) of the \code{Dotto}'s folder whom we want to make a family with. This can
be the Dotto's label ( e.g., \code{D001}) or the full address
(e.g., \code{path/to/D001-2020-11-21-foo}).}

\item{open}{If open the dotto file in Rstudio or not. Default is \code{TRUE}.}
}
\description{
It creates a template of \code{Dotto} based on an existing \code{Dotto} and we call them a family.
The new \code{Dotto} automatically has a line connection to the base \code{Dotto} and the card generated in
the index page combine the information related to all family.
}
