% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{other_docs}
\alias{other_docs}
\title{Title}
\usage{
other_docs(other = NULL)
}
\arguments{
\item{other}{The name of another R packages to build documentation for. Can only accept one package at a time. Must be a character vector, and equal to the folder name of the repository in the local GitHub folder.}
}
\description{
Title
}
