% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_docs.R
\name{other_docs}
\alias{other_docs}
\title{other_docs}
\usage{
other_docs(pkg = "/Users/evanodell/Documents/GitHub/",
  path = "/Users/evanodell/Documents/GitHub/docs/", other = NULL)
}
\arguments{
\item{pkg}{Location of files for input}

\item{path}{Output location, defaults to this folder.}

\item{other}{The name of another R packages to build documentation for.}
}
\description{
other_docs
}
