% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_docs.R
\name{other_docs}
\alias{other_docs}
\title{other_docs}
\usage{
other_docs(pkg = "/Users/evanodell/Documents/GitHub/", other = NULL)
}
\arguments{
\item{pkg}{Location of files for input}

\item{other}{The name of another R packages to build documentation for.}
}
\description{
Function to build documentation for other R packages not listed in 
\code{\link{build_docs}}.
}
