% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_dotto.R
\name{render_dotto}
\alias{render_dotto}
\title{Render Dotto with more relax path and remove unnecessary libs}
\usage{
render_dotto(path = NULL)
}
\arguments{
\item{path}{the path to the .Rmd file. If \code{NULL}, it automatically detect the .Rmd file in
the current work directory (in case there is only one .Rmd file).}
}
\value{

}
\description{
Render Dotto with more relax path and remove unnecessary libs
}
